Add new parser for handling various CBOR-LD versions.
