+++ lib/decode.js
@@ -7,6 +7,7 @@ import {CborldError} from './CborldError.js';
 import {Converter} from './Converter.js';
 import {Decompressor} from './Decompressor.js';
 import {inspect} from './util.js';
+import {parse} from './parser.js';
 import {default as varint} from 'varint';
 
 // 0xd9 == 11011001
@@ -53,29 +54,25 @@ export async function decode({
     throw new TypeError('"cborldBytes" must be a Uint8Array.');
   }
 
-  // 0xd9 == 11011001
-  // 110 = CBOR major type 6
-  // 11001 = 25, 16-bit tag size (65536 possible values)
-  if(cborldBytes[0] !== CBOR_TAG_FIRST_BYTE) {
-    throw new CborldError(
-      'ERR_NOT_CBORLD',
-      'CBOR-LD must start with a CBOR major type "Tag" header of `0xd9`.');
-  }
-  const {suffix, legacyFormat, registryEntryId} = _getSuffix({cborldBytes});
-  const isCompressed = _checkCompressionMode({cborldBytes, legacyFormat});
-  if(!isCompressed) {
-    return cborg.decode(suffix, {useMaps: false});
+  // parse CBOR-LD into a header and payload
+  const {header, payload} = parse({cborldBytes});
+  //console.log('header', header);
+  //console.log('payload', payload);
+
+  if(!header.payloadCompressed) {
+    return payload;
   }
 
   // decoded output is the abstract CBOR-LD input for conversion to JSON-LD
-  const input = cborg.decode(suffix, {useMaps: true});
+  const input = payload;//cborg.decode(suffix, {useMaps: true});
   if(diagnose) {
     diagnose('Diagnostic CBOR-LD decompression transform map(s):');
     diagnose(inspect(input, {depth: null, colors: true}));
   }
 
   // lookup typeTable by id if needed
-  if(legacyFormat !== 'singleton') {
+  const {format, registryEntryId} = header;
+  if(format !== 'legacy-singleton') {
     if(typeTable && typeTableLoader) {
       throw new TypeError('Use either "typeTable" or "typeTableLoader".');
     }
@@ -96,7 +93,7 @@ export async function decode({
   }
 
   const converter = _createConverter({
-    legacyFormat,
+    format,
     typeTable,
     documentLoader,
     appContextMap
@@ -111,12 +108,12 @@ export async function decode({
 
 // FIXME: consider making static method on `Converter`
 function _createConverter({
-  legacyFormat,
+  format,
   typeTable,
   documentLoader,
   appContextMap
 }) {
-  const legacyConverter = legacyFormat === 'singleton';
+  const legacyConverter = format === 'legacy-singleton';
   typeTable = legacyConverter ?
     createLegacyTypeTable({typeTable, appContextMap}) :
     createTypeTable({typeTable});
@@ -161,7 +158,6 @@ function _checkCompressionMode({cborldBytes, legacyFormat}) {
     }
     return true;
   }
-
 }
 
 function _getSuffix({cborldBytes}) {
